requirements.txt
aiohttp
aiosignal
async-timeout
attrs
certifi
charset-normalizer
discord-webhook
discord.py
ffmpeg-python
frozenlist
future
idna
imageio-ffmpeg
multidict
numpy
py-cord
pycord
requests
urllib3
yarl